% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{Clean data}
\usage{
eq_clean_data(data)
}
\arguments{
\item{data}{A data frame with raw data}
}
\value{
This function returns a dataframe with the information from the file after being cleaned.
}
\description{
This is a simple function that clean the data
}
\examples{
\dontrun{
data <- readr::read_delim("signif.tsv.gz", delim = "\\t")
eq_clean_data(data)
}

}
