% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{eq_create_label}
\alias{eq_create_label}
\title{Creates a label for leaflet map}
\usage{
eq_create_label(data)
}
\arguments{
\item{data}{A data frame with the earthquake data}
}
\value{
A vector with labels
}
\description{
This function takes the dataset as an argument and
 creates an HTML label that can be used as the annotation text in the leaflet map
}
\examples{
\dontrun{
eq_create_label(data)
}
}
