% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{Clean location}
\usage{
eq_location_clean(data)
}
\arguments{
\item{data}{A data frame with raw data}
}
\value{
This function returns a dataframe with a cleaned LOCATION_NAME column
}
\description{
This is a simple function that clean the column location in a dataframe
}
\examples{
\dontrun{
eq_location_clean(data)
}

}
