% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{eq_map}
\alias{eq_map}
\title{Map of earthquakes}
\usage{
eq_map(data, annot_col)
}
\arguments{
\item{data}{A data frame with the earthquake data}

\item{annot_col}{A string with the name of the column to show in the pop up}
}
\value{
A Map with earthquakes and annotations.
}
\description{
The function maps the epicenters (LATITUDE/LONGITUDE) and annotates
 each point with in pop up window containing annotation data stored
 in a column of the data frame
}
\examples{
\dontrun{
eq_map(data, annot_col = "Date")
}
}
