% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{geom_timeline}
\alias{geom_timeline}
\title{Time line of earthquakes}
\usage{
geom_timeline(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\description{
This geom for ggplot2 called geom_timeline() for plotting a time line
 of earthquakes ranging from xmin to xmaxdates with a point for each earthquake.
 Optional aesthetics include color, size, and alpha (for transparency)
}
\examples{
\dontrun{
data \%>\% eq_clean_data() \%>\%
dplyr::filter(Date >= "1980-01-01" & Date <="2014-01-01" & COUNTRY == c("ITALY","USA", "JORDAN")) \%>\%
  ggplot(aes(x = Date,
             y = COUNTRY,
             color = DEATHS,
             size = EQ_MAG_ML
  )) +
  geom_timeline() +
  theme_timeline()

}

}
