% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{theme_timeline}
\alias{theme_timeline}
\title{Theme for a correct visualization}
\usage{
theme_timeline()
}
\description{
This is a theme to configure the look of the geom
}
\examples{
\dontrun{
data \%>\% eq_clean_data() \%>\%
dplyr::filter(Date >= "1980-01-01" & Date <="2014-01-01" & COUNTRY == c("ITALY","USA", "JORDAN")) \%>\%
  ggplot(aes(x = Date,
             y = COUNTRY,
             color = DEATHS,
             size = EQ_MAG_ML
  )) +
  geom_timeline() +
  theme_timeline()

}

}
